+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## Improvements
 - PR #353 Change snmg python wrapper in accordance to cpp api
+- PR #365 Updates for setting device and vertex ids for snmg pagerank
 
 ## Bug Fixes
 - PR #354 Fixed bug in building a debug version